RequestDetailClientを作成
